♻️ refactor : deal and concert interfaces
