+++ frontend/craco.config.js
@@ -17,6 +17,7 @@ module.exports = {
       "@utils": path.resolve(__dirname, "src/utils"),
       "@hooks": path.resolve(__dirname, "src/hooks"),
       recoils: path.resolve(__dirname, "src/recoils"),
+      "@interfaces": path.resolve(__dirname, "src/interfaces"),
     },
   },
 };
