+++ frontend/src/interfaces/concert.interface.tsx
@@ -1,0 +1,20 @@
+export interface ConcertInterface {
+  imageUrl: string;
+  kopisConcertId: string;
+  place: string;
+  title: string;
+  startDate?: string;
+  startTime?: string;
+  concertId?: number;
+  longitude?: number;
+  latitude?: number;
+}
+
+// search modal
+// export interface ConcertInterface {
+//   imageUrl: string;
+//   kopisConcertId: string;
+//   place: string;
+//   startDate: string;
+//   title: string;
+// }
