+++ frontend/src/interfaces/deal.interface.tsx
@@ -1,0 +1,14 @@
+// DealCard
+export interface DealInfoInterface {
+  dealId: number;
+  dealTypeName: string;
+  endTime?: string;
+  imageUrl: string;
+  isBookmarked: boolean;
+  limitation?: number;
+  meetingLocation: string;
+  price?: number;
+  startTime: string;
+  title: string;
+}
+
